requirements.txt
flask==2.2.3
numpy==1.24.2
opencv-python==4.7.0.72
torch==2.0.0
transformers==4.27.4
openai-whisper==20230314
googletrans==4.0.0-rc1
gtts==2.3.1
mediapipe==0.10.9
Werkzeug==2.2.3
safetensors

